Moved util functions from script to modules
Minor extensions and simplifications beyond that
