+++ tianshou/utils/net/continuous.py
@@ -5,13 +5,31 @@ import numpy as np
 import torch
 from torch import nn
 
-from tianshou.utils.net.common import MLP
+from tianshou.utils.net.common import MLP, get_module_device
+from tianshou.utils.types import TDevice
 
 SIGMA_MIN = -20
 SIGMA_MAX = 2
 
 
-class Actor(nn.Module):
+class PreprocessNetWrapper(nn.Module):
+    def __init__(self, preprocess_net: nn.Module, device: TDevice = "cpu") -> None:
+        """
+        :param preprocess_net:
+        :param device: if None, use the same device as the preprocess_net.
+            In that case, will assume that all parameters of preprocess_net
+            are on the same device.
+        """
+        super().__init__()
+        device = device if device else get_module_device(preprocess_net)
+        preprocess_net.to(device)
+        self.device = device
+        self.preprocess_net = preprocess_net
+        # for backward compatibility
+        self.preprocess = preprocess_net
+
+
+class Actor(PreprocessNetWrapper):
     """Simple actor network. Will create an actor operated in continuous \
     action space with structure of preprocess_net ---> action_shape.
 
@@ -23,6 +41,9 @@ class Actor(nn.Module):
         only a single linear layer).
     :param float max_action: the scale for the final action logits. Default to
         1.
+    :param device: if None, use the same device as the preprocess_net.
+        In that case, will assume that all parameters of preprocess_net
+        are on the same device.
     :param int preprocess_net_output_dim: the output dimension of
         preprocess_net.
 
@@ -44,16 +65,11 @@ class Actor(nn.Module):
         device: Union[str, int, torch.device] = "cpu",
         preprocess_net_output_dim: Optional[int] = None,
     ) -> None:
-        super().__init__()
-        self.device = device
-        self.preprocess = preprocess_net
+        super().__init__(preprocess_net, device=device)
         self.output_dim = int(np.prod(action_shape))
         input_dim = getattr(preprocess_net, "output_dim", preprocess_net_output_dim)
         self.last = MLP(
-            input_dim,  # type: ignore
-            self.output_dim,
-            hidden_sizes,
-            device=self.device
+            input_dim, self.output_dim, hidden_sizes, device=self.device  # type: ignore
         )
         self.max_action = max_action
 
@@ -69,7 +85,7 @@ class Actor(nn.Module):
         return logits, hidden
 
 
-class Critic(nn.Module):
+class Critic(PreprocessNetWrapper):
     """Simple critic network. Will create an actor operated in continuous \
     action space with structure of preprocess_net ---> 1(q value).
 
@@ -78,7 +94,10 @@ class Critic(nn.Module):
     :param hidden_sizes: a sequence of int for constructing the MLP after
         preprocess_net. Default to empty sequence (where the MLP now contains
         only a single linear layer).
-    :param int preprocess_net_output_dim: the output dimension of
+    :param device: if None, use the same device as the preprocess_net.
+        In that case, will assume that all parameters of preprocess_net
+        are on the same device.
+    :param preprocess_net_output_dim: the output dimension of
         preprocess_net.
     :param linear_layer: use this module as linear layer. Default to nn.Linear.
     :param bool flatten_input: whether to flatten input data for the last layer.
@@ -97,14 +116,12 @@ class Critic(nn.Module):
         self,
         preprocess_net: nn.Module,
         hidden_sizes: Sequence[int] = (),
-        device: Union[str, int, torch.device] = "cpu",
+        device: TDevice = None,
         preprocess_net_output_dim: Optional[int] = None,
         linear_layer: Type[nn.Linear] = nn.Linear,
         flatten_input: bool = True,
     ) -> None:
-        super().__init__()
-        self.device = device
-        self.preprocess = preprocess_net
+        super().__init__(preprocess_net, device=device)
         self.output_dim = 1
         input_dim = getattr(preprocess_net, "output_dim", preprocess_net_output_dim)
         self.last = MLP(
@@ -123,24 +140,18 @@ class Critic(nn.Module):
         info: Dict[str, Any] = {},
     ) -> torch.Tensor:
         """Mapping: (s, a) -> logits -> Q(s, a)."""
-        obs = torch.as_tensor(
-            obs,
-            device=self.device,
-            dtype=torch.float32,
-        ).flatten(1)
+        obs = torch.as_tensor(obs, device=self.device, dtype=torch.float32).flatten(1)
         if act is not None:
-            act = torch.as_tensor(
-                act,
-                device=self.device,
-                dtype=torch.float32,
-            ).flatten(1)
+            act = torch.as_tensor(act, device=self.device, dtype=torch.float32).flatten(
+                1
+            )
             obs = torch.cat([obs, act], dim=1)
         logits, hidden = self.preprocess(obs)
         logits = self.last(logits)
         return logits
 
 
-class ActorProb(nn.Module):
+class ActorProb(PreprocessNetWrapper):
     """Simple actor network (output with a Gauss distribution).
 
     :param preprocess_net: a self-defined preprocess_net which output a
@@ -151,6 +162,9 @@ class ActorProb(nn.Module):
         only a single linear layer).
     :param float max_action: the scale for the final action logits. Default to
         1.
+    :param device: if None, use the same device as the preprocess_net.
+        In that case, will assume that all parameters of preprocess_net
+        are on the same device.
     :param bool unbounded: whether to apply tanh activation on final logits.
         Default to False.
     :param bool conditioned_sigma: True when sigma is calculated from the
@@ -178,21 +192,16 @@ class ActorProb(nn.Module):
         conditioned_sigma: bool = False,
         preprocess_net_output_dim: Optional[int] = None,
     ) -> None:
-        super().__init__()
+        super().__init__(preprocess_net, device=device)
         if unbounded and not np.isclose(max_action, 1.0):
             warnings.warn(
                 "Note that max_action input will be discarded when unbounded is True."
             )
             max_action = 1.0
-        self.preprocess = preprocess_net
-        self.device = device
         self.output_dim = int(np.prod(action_shape))
         input_dim = getattr(preprocess_net, "output_dim", preprocess_net_output_dim)
         self.mu = MLP(
-            input_dim,  # type: ignore
-            self.output_dim,
-            hidden_sizes,
-            device=self.device
+            input_dim, self.output_dim, hidden_sizes, device=self.device  # type: ignore
         )
         self._c_sigma = conditioned_sigma
         if conditioned_sigma:
@@ -200,7 +209,7 @@ class ActorProb(nn.Module):
                 input_dim,  # type: ignore
                 self.output_dim,
                 hidden_sizes,
-                device=self.device
+                device=self.device,
             )
         else:
             self.sigma_param = nn.Parameter(torch.zeros(self.output_dim, 1))
@@ -275,11 +284,7 @@ class RecurrentActorProb(nn.Module):
         info: Dict[str, Any] = {},
     ) -> Tuple[Tuple[torch.Tensor, torch.Tensor], Dict[str, torch.Tensor]]:
         """Almost the same as :class:`~tianshou.utils.net.common.Recurrent`."""
-        obs = torch.as_tensor(
-            obs,
-            device=self.device,
-            dtype=torch.float32,
-        )
+        obs = torch.as_tensor(obs, device=self.device, dtype=torch.float32)
         # obs [bsz, len, dim] (training) or [bsz, dim] (evaluation)
         # In short, the tensor's shape in training phase is longer than which
         # in evaluation phase.
@@ -292,10 +297,11 @@ class RecurrentActorProb(nn.Module):
             # we store the stack data in [bsz, len, ...] format
             # but pytorch rnn needs [len, bsz, ...]
             obs, (hidden, cell) = self.nn(
-                obs, (
+                obs,
+                (
                     state["hidden"].transpose(0, 1).contiguous(),
-                    state["cell"].transpose(0, 1).contiguous()
-                )
+                    state["cell"].transpose(0, 1).contiguous(),
+                ),
             )
         logits = obs[:, -1]
         mu = self.mu(logits)
@@ -310,7 +316,7 @@ class RecurrentActorProb(nn.Module):
         # please ensure the first dim is batch size: [bsz, len, ...]
         return (mu, sigma), {
             "hidden": hidden.transpose(0, 1).detach(),
-            "cell": cell.transpose(0, 1).detach()
+            "cell": cell.transpose(0, 1).detach(),
         }
 
 
@@ -348,11 +354,7 @@ class RecurrentCritic(nn.Module):
         info: Dict[str, Any] = {},
     ) -> torch.Tensor:
         """Almost the same as :class:`~tianshou.utils.net.common.Recurrent`."""
-        obs = torch.as_tensor(
-            obs,
-            device=self.device,
-            dtype=torch.float32,
-        )
+        obs = torch.as_tensor(obs, device=self.device, dtype=torch.float32)
         # obs [bsz, len, dim] (training) or [bsz, dim] (evaluation)
         # In short, the tensor's shape in training phase is longer than which
         # in evaluation phase.
@@ -361,17 +363,13 @@ class RecurrentCritic(nn.Module):
         obs, (hidden, cell) = self.nn(obs)
         obs = obs[:, -1]
         if act is not None:
-            act = torch.as_tensor(
-                act,
-                device=self.device,
-                dtype=torch.float32,
-            )
+            act = torch.as_tensor(act, device=self.device, dtype=torch.float32)
             obs = torch.cat([obs, act], dim=1)
         obs = self.fc2(obs)
         return obs
 
 
-class Perturbation(nn.Module):
+class Perturbation(PreprocessNetWrapper):
     """Implementation of perturbation network in BCQ algorithm. Given a state and \
     action, it can generate perturbed action.
 
@@ -395,12 +393,10 @@ class Perturbation(nn.Module):
         preprocess_net: nn.Module,
         max_action: float,
         device: Union[str, int, torch.device] = "cpu",
-        phi: float = 0.05
+        phi: float = 0.05,
     ):
         # preprocess_net: input_dim=state_dim+action_dim, output_dim=action_dim
-        super(Perturbation, self).__init__()
-        self.preprocess_net = preprocess_net
-        self.device = device
+        super().__init__(preprocess_net, device=device)
         self.max_action = max_action
         self.phi = phi
 
@@ -442,7 +438,7 @@ class VAE(nn.Module):
         hidden_dim: int,
         latent_dim: int,
         max_action: float,
-        device: Union[str, torch.device] = "cpu"
+        device: Union[str, torch.device] = "cpu",
     ):
         super(VAE, self).__init__()
         self.encoder = encoder
@@ -475,17 +471,19 @@ class VAE(nn.Module):
         return reconstruction, mean, std
 
     def decode(
-        self,
-        state: torch.Tensor,
-        latent_z: Union[torch.Tensor, None] = None
+        self, state: torch.Tensor, latent_z: Union[torch.Tensor, None] = None
     ) -> torch.Tensor:
         # decode(state) -> action
         if latent_z is None:
             # state.shape[0] may be batch_size
             # latent vector clipped to [-0.5, 0.5]
-            latent_z = torch.randn(state.shape[:-1] + (self.latent_dim, )) \
-                .to(self.device).clamp(-0.5, 0.5)
+            latent_z = (
+                torch.randn(state.shape[:-1] + (self.latent_dim,))
+                .to(self.device)
+                .clamp(-0.5, 0.5)
+            )
 
         # decode z with state!
-        return self.max_action * \
-            torch.tanh(self.decoder(torch.cat([state, latent_z], -1)))
+        return self.max_action * torch.tanh(
+            self.decoder(torch.cat([state, latent_z], -1))
+        )
